Deprecated methods implemented.
Traverse all units in each block.
For Jimple, a Unit is a Stmt.
May check if it contains an invocation and retrieve it.
The invocation is an InvokeExpr :: Expr. It has a reference to the
defining method m_def.

To check if m_def is deprecated, check traverse Tag(s) and check if
there is an instance of a VisabilityAnnotationTag with type "Ljava/lang/Deprecated;".